refactor: split email queue tasks
